Call get_suggestions before `ignore word under cursor` for correct work
With spellers where `ignore word` is supported by speller library itself
we need to call get_suggestions because in case we use multiple languages
we need to ignore a word by the speller selected by the logic
of this function. This is not ideal but this functionality is not the
most important so I think it's ok.

#182